requirements.txt
Django
djangorestframework
markdown
psycopg2
coverage
